requirements.txt
streamlit>=1.28.0
langchain
langchain-google-genai
langchain-community
pypdf
chromadb
langchain-chroma
google-generativeai
python-dotenv
firebase-admin
requests
pytest>=7.0.0
google-auth
google-cloud-storage